[change]経験値に関わる引数を全てlong型に

+++ src/main/java/hugu1026/com/github/phantasystatus/command/GiveExpCommand.java
@@ -2,7 +2,6 @@ package hugu1026.com.github.phantasystatus.command;
 
 import hugu1026.com.github.phantasystatus.PhantasyStatus;
 import hugu1026.com.github.phantasystatus.event.GetExpEvent;
-import hugu1026.com.github.phantasystatus.util.PlayerDataUtil;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
@@ -22,7 +21,7 @@ public class GiveExpCommand implements CommandExecutor {
             if (!(sender instanceof Player)) {
                 if (args[0] == null || args[1] == null) return false;
                 Player player = Bukkit.getPlayer(args[0]);
-                int amount = Integer.parseInt(args[1]);
+                long amount = Long.parseLong(args[1]);
                 GetExpEvent event = new GetExpEvent(player, amount);
                 Bukkit.getServer().getPluginManager().callEvent(event);
             }
